+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Nested Step finish refactoring
+### Added
+- More JSR-305 annotations
 
 ## [5.1.0-RC-11]
 ### Changed